Fixed issues after Usertest 1 #84 #88 #91 #87 #89 #90
Added arrow in signupview so we can switch from signup to login. Changed the password in signup from TextField to PasswordField. Added togglegroup for mann/kvinne. Added "feedback" field in signupmodel, controller and view to give feedback for invalid user-input. Changed the placement of the textfields and the arrow in signupview.