First commit HomeScreenView and HomeScreenController#95
Created homeScreenView, and HomeScreenController, neither of them 100%
finished. The WebView in HomeScreenView shows interactive google map,
created with help of main.css, main.js, location.js and location.html.
Used leaflet and geoJSON. Also committed is the FxAppTest, which is
created for running the view. Will be deleted before branching to
master, but needed for testing.