Created LoggedInMVC with tests #66
Created test for LoggedInModel and LoggedInController. Used SceneBuilder to create the LoggedIn-view. Created LoggedInModel, with methods to get name and workouts for a user,  and LoggedInController, that connects the view with the model and the FxApp.

Added a maxpulse-field in Workout. Added tests in WorkoutTest to test the maxpulse-field.

Added path to LoggedIn-view in FxApp. Added field for coach(boolean) in FxApp. Changed paths in LoginController, SignupController and WorkoutRegistrationController, to lead to the LoggedIn-view.