Finished LoggedIn MVC #66 #68 #61
Changed endpoint in getPulsezones in LoggedInModel to work with backend. Added method "stringToDate()" to handle the dates from backend. Added tests for this method.

Changed the setChart() method in LoggedInController to handle both dates and duration in pulsezones from backend. Did this by adding two methods: getDay() and getDate(). Added tests for this method.

Added an extra field, dateString, in Workout that was needed to handle the date from backend in a better way.

I also added more comments to the code.